CCM-7385: Add dev tests
