+++ tests/development/message_batches/create_message_batches/test_success.py
@@ -172,38 +172,3 @@ def test_request_without_dob(nhsd_apim_proxy_url, bearer_token_internal_dev):
         data["data"]["attributes"]["messageBatchReference"],
         data["data"]["attributes"]["routingPlanId"],
     )
-
-
-@pytest.mark.devtest
-def test_201_message_batches_request_idempotency(
-    nhsd_apim_proxy_url,
-    bearer_token_internal_dev
-):
-    """
-    .. include:: ../../partials/happy_path/test_201_message_batches_request_idempotency.rst
-    """
-    data = Generators.generate_valid_create_message_batch_body("dev")
-
-    resp_one = requests.post(
-        f"{nhsd_apim_proxy_url}{MESSAGE_BATCHES_ENDPOINT}",
-        headers={
-            "Authorization": bearer_token_internal_dev.value,
-            "Accept": constants.DEFAULT_CONTENT_TYPE,
-            "Content-Type": constants.DEFAULT_CONTENT_TYPE,
-        },
-        json=data,
-    )
-
-    time.sleep(5)
-
-    resp_two = requests.post(
-        f"{nhsd_apim_proxy_url}{MESSAGE_BATCHES_ENDPOINT}",
-        headers={
-            "Authorization": bearer_token_internal_dev.value,
-            "Accept": constants.DEFAULT_CONTENT_TYPE,
-            "Content-Type": constants.DEFAULT_CONTENT_TYPE,
-        },
-        json=data,
-    )
-
-    Assertions.assert_message_batches_idempotency(resp_one, resp_two)
