+++ tests/development/messages/create_messages/test_success.py
@@ -120,29 +120,3 @@ def test_request_without_dob(nhsd_apim_proxy_url, bearer_token_internal_dev):
     )
 
     Assertions.assert_201_response_messages(resp, nhsd_apim_proxy_url)
-
-
-@pytest.mark.devtest
-def test_201_message_request_idempotency(nhsd_apim_proxy_url, bearer_token_internal_dev):
-    """
-    .. include:: ../../partials/happy_path/test_201_messages_request_idempotency.rst
-    """
-    data = Generators.generate_valid_create_message_body("dev")
-
-    resp_one = requests.post(f"{nhsd_apim_proxy_url}{MESSAGES_ENDPOINT}", headers={
-            "Authorization": bearer_token_internal_dev.value,
-            "Accept": constants.DEFAULT_CONTENT_TYPE,
-            "Content-Type": constants.DEFAULT_CONTENT_TYPE
-        }, json=data
-    )
-
-    time.sleep(5)
-
-    resp_two = requests.post(f"{nhsd_apim_proxy_url}{MESSAGES_ENDPOINT}", headers={
-            "Authorization": bearer_token_internal_dev.value,
-            "Accept": constants.DEFAULT_CONTENT_TYPE,
-            "Content-Type": constants.DEFAULT_CONTENT_TYPE
-        }, json=data
-    )
-
-    Assertions.assert_messages_idempotency(resp_one, resp_two)
